requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b4nang6w_y/croot/cffi_1700254307954/work
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
cvxopt==1.3.2
cycler==0.12.1
cymem==2.0.8
datasets==2.18.0
detect-secrets==1.4.0
dill==0.3.8
evaluate==0.4.1
Faker==23.3.0
filelock @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d3quwmvouf/croot/filelock_1700591194006/work
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_c7sublivu7/croot/fsspec_1701286493452/work
ftfy==6.2.0
fuzzysearch==0.7.3
gast==0.5.4
gmpy2 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/croot-muov7gw1/gmpy2_1645462616268/work
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
hmmlearn @ git+https://github.com/hmmlearn/hmmlearn@23c0f132bd66280b46b7f898dc51812629c8bdb7
huggingface-hub==0.22.1
idna==3.6
install==1.3.5
Jinja2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_7dognxkzoy/croot/jinja2_1706733627811/work
joblib==1.3.2
json_repair==0.9.0
keras==3.1.1
kiwisolver==1.4.5
langcodes==3.3.0
libclang==18.1.1
lightning-utilities @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_decp_2qeye/croot/lightning-utilities_1694002425245/work
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a84ni4pci8/croot/markupsafe_1704206002077/work
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_17iu6a8a3m/croot/mpmath_1690848269369/work
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.7
networkx @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b9af3smw_7/croot/networkx_1690562010704/work
nltk==3.8.1
numpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a51i_mbs7m/croot/numpy_and_numpy_base_1708638620867/work/dist/numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl#sha256=c4b11b3c4d4fdb810039503fe01f311ade06cd1d675fcd6d208800a393f19b69
openai-clip==1.0.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.10.0
packaging @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a6lqg7at4g/croot/packaging_1710807410750/work
pandas==2.2.1
peft==0.10.0
phonenumbers==8.13.33
pillow==10.2.0
preshed==3.0.9
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
responses==0.18.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
seqeval==1.2.2
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
span-marker==1.5.0
srsly==2.4.8
structlog==24.1.0
sympy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_7cbpg8656h/croot/sympy_1701397648473/work
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras==2.16.0
thinc==8.2.3
threadpoolctl==3.2.0
tiktoken==0.6.0
tldextract==5.1.2
tokenizers==0.15.2
torch==2.1.2
torcheval==0.0.7
torchmetrics @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_11rlk64ar0/croot/torchmetrics_1701875357241/work
torchvision==0.16.2
tqdm==4.66.2
transformers==4.38.2
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
